Feature/a1 mtdc@148 edc system test (#248)
Added a system test for the connector deployment.

Because the file copy inputs and outputs are stored on the Provider, the test temporarily uses kubectl to access files. This will be changed later when we integrate with PRS API and blob storage.
For the same reason, the Consumer can't poll the storage to detect job completion. Therefore the test is currently not covering the status polling endpoint.