Feature/a1 mtdc@171 edc job (#280)
The logic to orchestate multiple transfers (to build a PRS parts tree out of recursive calls to PRS API endpoints in multiple dataspace partitions) is factored out into a generic framework to manage jobs, that allows plugging in a custom handler to provide the logic, in our case to parse the output of one PRS API call to determine the next endpoints to call.

This PR is only the job framework. The custom logic is as before, and runs only a single PRS API call. Further PRs willl build on this to run recursive PRS API calls.

Note that jobs accumulate in in-memory storage. Created follow-up task 179 to deal with memory management since this could be a broader topic including blob storage cleanup.