Feature/a1 mtdc@165 blob fix (#297)
Some fixes for the previous PRs on blob storage and distributed PRS:

Mount a directory rather than a file in docker-compose. If mounting a file directly, if the file doesn't exist previously, docker-compose creates a directory at that location which blocks file download, it also messes up the image and makes fixing this a pain.
Use a proper One ID in docker-compose tests.
Call complete() handler if no transfers are run.